requirements.txt
conda:
    torch>=1.1
    torchvision
    opencv
    scikit-image
    matplotlib
    tensorboard>=1.14
    pandas
    tqdmp

pip:
    hydra-core>=1.1
    hydra_colorlog
    einops
    kornia
    pytorch-msssim
    ptflops
    pyiqa
    requests
    pytorch-gradual-warmup-lr